Added a disambiguation step to function resolution that says functions with where clauses should be favored over functions without where clauses.
Updated the spec.  Also, fixed a mistake in the written algorithm for
determining more specific functions in the spec in which we implied
that a better argument would be used before the lack of promotion in
determining that a function should be favored by function resolution.
We favor functions that do not promote even if some or all of the
arguments are worse matches.  Made the algorithm a bit more exact as
well.

Added the new disambiguation step to the specification.

There is an interesting issue that showed up via the test suite when
making this change.  It regards whether we should favor a function
with a where clause over a function that shadows it or vice versa.
Before delving into it, I'll include some background from e-mail
discussions that provides motivation for this change.  Search ahead
for "Issue" to skip this background.

Background:
----------

Below I've excerpted my original summary e-mail that provides
motivation for this change and Brad's response that adds further
justification.  I also added a note to my e-mail that points out the
fix to the written algorithm in the specification.

I wrote:

Here is an example that motivates this change.  Suppose we want to
write something like

   def foo(A: []) where A.rank == 1 && A.eltType == 3*int { ... }
   def foo(A: []) where A.rank == 2 && A.eltType == 2*int { ... }
   def foo(A: []) where A.rank == 3 && A.eltType == int { ... }
   def foo(A: []) { ... }

where the fourth overload applies only if the other three do not.
Currently if we call foo such that one of the first three overloads
applies, then we get an ambiguity.  For example,

   var A: [1..3] 3*int; foo(A); // ambiguity between 1st and 4th

So we actually have to write the fourth function as

   def foo(A: []) where !(A.rank == 1 && A.eltType == 3*int) &&
                        !(A.rank == 2 && A.eltType == 2*int) &&
                        !(A.rank == 3 && A.eltType == int) { ... }

so that the fourth overload no longer applies if any of the first
three overloads apply.

In order to support the omission of these where clauses we have to
change the function resolution algorithm for determining whether a
given function is better than another function.  This algorithm is
explained in the spec as below where I've added a line (marked by *)
to allow us to choose a function with a where clause over one without.

   Given two functions F1 and F2, F1 is determined to be more specific
   than F2 by the following steps:

     If F1 does not require promotion and F2 does require promotion,
     then F1 is more specific.

[N.B. The bug I refer to in the specification in this commit was that
 the previous statement used to be below the following statement.]

     If at least one of the legal argument mappings to F1 is a more
     specific argument mapping than the corresponding legal argument
     mapping to F2 and none of the legal argument mappings to F2 is a
     more specific argument mapping than the corresponding legal
     argument mapping to F1, then F1 is more specific.

     If F1 shadows F2, then F1 is more specific.

*   If F1 has a where clause and F2 does not have a where clause, then
*   F1 is more specific.

     Otherwise, F1 is not more specific than F2.

I'm in favor of this change based on the intuitive argument of the
common case: Given the example above, this change will make it easier
to write the code that programmers are likely to want to write.

I'm against this change based on the complexity argument: Function
resolution already has too many steps, making it very difficult to
reason about which function will apply in some cases.

Here is an odd example where we don't get what we want, I would argue.
Imagine function overloads

   def foo(type t, x: t, y: real) where t == int { ... }
   def foo(type t, x: real, y: t)                { ... }

and a call

   foo(int, 2, 2);

Without the where clause, we have an ambiguity because the second
actual argument matches the first overload's second formal argument
better than the second overload's second formal argument but the third
actual argument matches the second overload's third formal argument
better than the first overload's third formal argument.  However, due
to the unrelated where clause, we favor the second overload and break
this ambiguity.  This is contrived, but I think there are real cases
that could show up and cause real confusion.

Given that I'm for and against making this change, I'll say that I'm
leaning towards making it in balance.  We already have some odd things
(like the promotion clause above) that help in natural ways.  This
addition I think helps in natural ways.  However, I would believe
there are other examples out there, contrived or not, that could
suggest this is a bad idea.  While it's unfortunate to have to add a
where clause that is just the negation of the other where clauses,
it's not all that awful and the alternative is probably clearer.

Brad added:

I think the main downside to maintaining the status quo is that in
order for it to be effective, you have to have control of all
overloads of a function.  That is, I can only write an overload that
serves as a "fall-through" or default case if I know what all of the
other overloads of the function are so that I can write a where clause
that is the exclusion of all other where clauses.  This may be
reasonable when all the overloads are located within one module and
controlled by one user, but if you have a function that several
modules may provide overloads for (where not all of those modules may
be used for any given compilation) then you don't have any way to
create the version of the function to be used when no other where
clauses apply.

Issue:
-----

There is a question as to whether we should favor a function without a
where clause over a function with a where clause if the function
without a where clause shadows the function with a where clause.  We
currently use shadowing first so we would.  This is important for the
test

  test/distributions/bradc/foundations/blockdist1.chpl

which fails otehrwise because a more visible function without a where
clause shadows several functions with where clauses.  This test
defines a function "def min(type t)" without a where clause.  Since
the standard functions with the same signature in the Types module
have where clauses, they were favored over this one which shadows
them.  If we deal with visiblity first, then we continue to favor the
test's function.

I think this is the right position for this step especially if we
think of the function without a where clause as an "otherwise" case.
In this "otherwise" purpose, it should probably be the shadowed
function as well.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16503 3a8e244f-b0f2-452b-bcba-4c88e055c3ca