Implement CHPL_TASK_SLEEP().
Add CHPL_TASK_SLEEP() to the tasking interface, and implement
Time.sleep() in terms of that, instead of calling the system's sleep()
routine directly.  This allows a tasking layer to do other things while
the task sleeps.  For example, a multiplexing tasking layer could
suspend the sleeping task for the duration of the sleep, and run other
tasks in the meantime.

We are discussing whether or not sleeping is representative of a more
general class of situations in which we might want to suspend or switch
tasks, and if so how that should be done.  So, this may change again in
the future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16504 3a8e244f-b0f2-452b-bcba-4c88e055c3ca