Add the codes in doc/tutorial/solutions to the testing system so we don't get surprised at the last minute if they stop working.
Added an explicit seed for the random numbers in the histogram[1-3] and life
solutions.

Used prediff scripts to normalize the output of intentionally nondeterministic
programs and filter out timers that didn't have a config var to turn them off.

EstimatePi7 is nondeterministic unintentionally and could get a different
answer from what it expects.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16506 3a8e244f-b0f2-452b-bcba-4c88e055c3ca