Cleaned the function disambiguate_by_match in function resolution. Reordered some conditionals to better match the specification (though there is no semantic change here.
Removed one pair of conditionals that used to determine that a type
argument was better than a non-type argument.  This comparison is no
longer necessary since we no longer allow values to be passed to type
arguments.

Futurized (semantic futures) two tests of bool coercions.  These used
to give ambiguous errors, but now they resolve to the first function
that is listed in the program.  This is because there is a circularity
in coercions (bool to bool(32) and bool(32) to bool, for example).

Renamed a few variables to make the algorithm much clearer.  For
example, I changed the name of the variable 'better' to 'worse' which
matches the natural ordering of the algorithm.  Since best refers to
whether fn1 is the best match, then when we compare it against fn2, it
makes more sense to ask whether fn1 is worse than fn2 rather than
whether fn2 is better than fn1.  Added comments.

Cleaned the algorithm in the spec, adding a few steps and removing
others so that it matches the algorithm in the implementation.  Most
of this cleaning is superficial and makes all of the steps more
symmetric with one another.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16507 3a8e244f-b0f2-452b-bcba-4c88e055c3ca