Renaming HPCC modules to avoid requirement to name them on chpl command line
This is a small housecleaning change I made while on the road.  Now
that we have a basic module search capability, it's much more
convenient to have the HPCC helper modules (originally probSize.chpl
and ra-randstream.chpl) use their module names as their filenames
(HPCCProblemSize and RARandomStream) to simplify the chpl command
lines when doing development work with them.  This permits most of the
.compopts files to be removed (ra still uses --no-warnings and -O) and
the Makefile/README instructions to be simplified.

Note that as currently written, these files still name their module
using a module NAME { ... } declaration within the code files.  This
isn't strictly necessary, but can be helpful if someone is going to
supply multiple implementations (e.g., implement their own random
stream for RA).  If people think this benefit is not worth the extra
lines of code (and resulting indentation), we could remove them.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16514 3a8e244f-b0f2-452b-bcba-4c88e055c3ca