Similar directory reuse change
For testing the release, we've been creating a unique directory
based on the pid of the cron job.  The release testing blows this
directory away, except for when failures occur, which is why this
isn't typically a problem.  In this change, I have the release
testing reuse the same directory to avoid this problem.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16523 3a8e244f-b0f2-452b-bcba-4c88e055c3ca