Removing explicit mention of renamed module files from ra test
This test piggy-backed on our test/release/examples/hpcc files
which were renamed late last year, but because this test only
fires on GASNet, I didn't catch this before leaving for vacation.
Replaced explicit naming of the module files in the compopts
file with a -M option that should do the trick.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16525 3a8e244f-b0f2-452b-bcba-4c88e055c3ca