Removed enumerated arrays/domains and added a special case in associative to handle them instead.
Added domain methods add, remove, and clear that issue compiler errors if they
are called on a domain with enumerated type indices.

Added an associative tupleInit method so associative arrays can be initialized
with tuples.

Added "clear" to the list of methods supported by associative domains but not
enumerated domains in the spec.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16527 3a8e244f-b0f2-452b-bcba-4c88e055c3ca