Adding initial support for external C struct types
This check-in has very simple and minimal support for declaring
external C struct types using _extern record { ... } in Chapel.  I've
documented this capability in its current form in the README.extern
file.

externFloat4.chpl exhibits this capability in a simple test case using
float4.h as the C file defining the external structure.
someFields.chpl shows that a subset of the complete set of fields can
be declared in the extern record declaration.

This version assumes that the external struct can be named in the C
code using its identifier "foo" (as opposed to "struct foo").
Eventually we will want to support both forms, as captured in
externFloat4b.chpl.  Its .future file captures my current thoughts on
how to do this.

Currently extern structs can't really be passed to/from external
routines because of the way we modify such calls to pass by pointer,
deal with return values, etc.  These issues are captured in
externFloat4calls.chpl and externFloat4Calls.someFields.chpl.

In order for this capability to be truly useful, I think we'll also
want to add a CTypes module which defines standard C types like int,
float, double, etc. in a way that they map to the appropriate Chapel
types.  Currently, the user has to handle the burden of choosing the
best matching Chapel type (as they do for external C routines as
well).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16530 3a8e244f-b0f2-452b-bcba-4c88e055c3ca