Removed primitives used to define C math functions (e.g., sin())
Removed all primitives that were used to define C math functions like
sin(), cos(), etc. and switched them to use _extern calls instead, the
argument being that the compiler doesn't need any special knowledge of
these calls.  Note that the Chapel versions of the functions are still
required in general because we support overloading (e.g., sin() works
for real(32) and real(64), rather than using sin() and sinf()).
However, as I type this, I'm realizing that most of the real(64) cases
could probably simply be standalone extern declarations.  I'll try to
make that change next.

I put the extern definitions of the C functions into the bodies of the
Chapel versions of the functions in order to prevent bleeding external
C functions into the Chapel space (e.g., sinf()) and checked in a test
to lock this in.  If module use was not transitive (and/or we had a
notion of "private" symbols), we could probably do this more
elegantly; possibly even making the C names optionally available to
programmers who preferred them.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16534 3a8e244f-b0f2-452b-bcba-4c88e055c3ca