Fixed a bug with the --no-codegen flag
It turned out that the (developer-oriented) --no-codegen
flag did not cause the buildBinary pass to be skipped and
therefore hasn't been working as intended for awhile.  This
commit re-enables it and adds a test to ensure that it's
working.

I think that over time, we may want to make --no-codegen
into a user flag and add some other flags that short-circuit
compilation at certain points such as --parse-only, --check-only
(or something that implies an exit after semantic checking),
--no-codegen, and --no-build-binary.  I didn't take that on here.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16545 3a8e244f-b0f2-452b-bcba-4c88e055c3ca