Started work to create a set of robustness tests for distributed arithmetic arrays.  The idea here is to use a single file driver.chpl that creates different types of distributions based on a configuration parameter.  For example, I set this up to work with Block by setting the parameter distString to 'block'.  In this case, we create four distributions (a 1D one, a 2D one, a 3D one, and a 4D one) which are then used by all of the tests in this directory to test the functionality of Block-distributed arrays and domains.  I added one test that makes sure that the elements of an array are initialized to 0 and that the values read match what is written.
I'm currently starting this out with four distributions, one for each
of the first four ranks, each with integral index type, and each
focused on the indices in rank*(1..5).  All distributions need to work
with all indices, but the distributions are meant to be defined so as
to distributed the indices in these ranges for these first four
distributions.  We can add more distributions to deal with other
cases.  For example, large int(64) indices, unsigned integral indices,
negative indices, higher dimensions, etc.

Unfortunately, I had to put the 'use' statement top-level in the file.
I tried to put the use statement inside the setupDistributions
function, but this didn't work because the functions are not
accessible outside.  The distribution, array, and domain records
escape the scope where we can do anything with them.  This setup also
thus requires transitive uses because the use statements are only
going to be in the driver.chpl files.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16546 3a8e244f-b0f2-452b-bcba-4c88e055c3ca