Continued work to setup a distribution robustness testing ground.
Added 'default' case to test the default distribution.  This is now
the default.  To test the Block distribution, use -sdistString='block'
as a compiler flag via -compopts.

Changed the bounds the tests and distributions should be concerned
about distributing and added constants to make it easier to write
tests.  For example, for Block, these bounds specify the bounding box.
These are specified as:

1D: const n1 = 100, Space1 = [1..n1]
2D: const n2 = 20, Space2 = [1..n2, 1..n2]
3D: const n3 = 5, Space3 = [1..n3, 1..n3, 1..n3]
4D: const n4 = 3, Space4 = [1..n4, 1..n4, 1..n4, 1..n4]

Rewrote the only test to use these constants.

Added a file with notes about what functionality we need to write
tests to exercise.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16547 3a8e244f-b0f2-452b-bcba-4c88e055c3ca