Removed most primitives from Memory.chpl
Replaced most of the primitives in Memory.chpl with extern calls.
The one primitive that remained was one which took filename/lineno
information.  In order to get rid of these primitives, we may want
to introduce some built-in cpp-style FILE/LINE parameters that expand
to the equivalent of what we pass around in insertLineNumbers rather
than always literally the line number in question (which is to say,
when used within internal modules (and possibly standard modules?),
they would reflect user line numbers.  This doesn't seem hard, but
I'm not taking it on here (or in the near future I don't think).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16548 3a8e244f-b0f2-452b-bcba-4c88e055c3ca