Fixed max int range failure and removed future.
This fix keeps the iterator simple (a yield in a loop, and a single
condition, but computing what the index will be when the loop is
finished.  This is arguably an improvement in the potentially strided
case because the condition low <= i && i <= high has been replaced
with i != last.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16551 3a8e244f-b0f2-452b-bcba-4c88e055c3ca