Make iteration over enumerated associative domains and arrays follow the order the enumeration constants were declared in. Changed the associative tupleInit() method to only be allowed for enumerated associative arrays.
Made the writers for associative arrays and domains use the default iterators
to define the order they output in. Since the iterators for enumerated
associative arrays/domains have a set order, this orders their output.

Added _isEnumeratedType(type t) to the _is*Type functions in ChapelBase.

Added chpl_enumerate(type t) iterator that takes an enumerated type and
yields the values to the Types module.

Added a test that walks over enumerated associative arrays in different ways
to ensure that they always follow the right order.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16572 3a8e244f-b0f2-452b-bcba-4c88e055c3ca