Rewrote the grep scripts to use find.  It's faster than having the shell expand the wildcards (sometimes as much as 4X, but typically more like 50%) and less fragile (if we change our directory structure, we don't have to change the scripts).
For the most part, I kept the behavior the same, but there is one
change.  The script do not automatically pass the -e flag to grep.
The benefit of this is that now you can pass arbitrary options to grep
to these scripts (for example, -i or -c or -l).  So the new usage is:

    grepchpl|greprt|grepmod|grepmake [grep option] PATTERN

I also removed the special casing for SunOS, because it didn't work
anyways (still doesn't work).

I was not able to test on the mac because it is not plugged into the
network.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16582 3a8e244f-b0f2-452b-bcba-4c88e055c3ca