Another small change to nightly test script
Made it so that when run in -debug mode, the default revision used is
the current one rather than that day's at midnight, as with -cron.
Also added support for a -r flag which allows a revision to be
specified which is passed along to the svn checkout/update commands.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16585 3a8e244f-b0f2-452b-bcba-4c88e055c3ca