Added support for new distribution syntax where 'new dist' replaces the placeholder syntax 'distributionValue' and 'dist' replaces the placeholder syntax 'distributionType'.  Extended syntactic support for the 'distributed' keyword to apply to domain expressions.
Updated syntax in the modules and tests.

I added a new flag "syntactic distribution" that applies to a new
record in the standard modules called 'dist'.  This lets the compiler
transform constructor calls and types related to this record into
function calls.

Note that there is more work related to these syntax changes.  I have
to add support for the syntactic sugar.  I have to enable assignment
over distributions since declaring a distribution type and then
initializing it with a distribution value doesn't yet work because
assignment is not allowed.  I've also noticed that our yacc file isn't
as general as it should be.  For example, you can write

  type t = int, tt = real;

in a function, but not in a class.  There are numerous other things
and, related to this change, domain literals with a distribution are
more restrictied than domains literals without a distribution as far
as where they can be written in a program syntax.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16587 3a8e244f-b0f2-452b-bcba-4c88e055c3ca