Continued reorganization of parser, including a more major cleanup and a few semantic changes.  Added the new operator to the precedence table (above : and below ., similar to C++) and relaxed its parsing. Allowed domain types, subdomain types, sparse domain types, and index types to be parsed wherever expressions are parsed.
In allowing domain types to be parsed wherever, distributed acts as an
operator on a non-distributed domain type.  I was forced to add a 'no
auto destroy' pragma on the domain that is built from this domain type
for sparse domains.  A comment explains this issue and I think it
indicates a bug with reference counting and runtime types.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16596 3a8e244f-b0f2-452b-bcba-4c88e055c3ca