Add the prefix "dsi" (distribution standard interface) to the functions that are the interface between the distribution records and classes.
Changed the accessor for domain and array classes from "this" to "dsiAccess".
Changed the writers for these classes from "writeThis" to "dsiSerialWrite".

Removed the dsi prefix from some array/domain wrapper record functions where
they were incorrectly added when we were discussing this change.

Did not change the default iterators from "these" to "dsiDefaultIterator".
The _getIterator function expects to be able to call these on multiple types,
not just arrays and domains. I plan to change this in another commit.

Updated tests that use or define the dsi functions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16597 3a8e244f-b0f2-452b-bcba-4c88e055c3ca