Implemented dsiBuildArithmeticDom() for Block distributions, which is neceessary for _domain methods expand, exterior, interior, and translate (and chpl__unTranslate).  Also cleaned up the implementation of these methods to avoid using _value if not necessary and also to do reference counting on the distribution class.
Also, moved dsiBuildArithmeticDom to the domain class (for Block and
Default) since it made more sense there.

Added 'dsi' prefix to a couple of methods in Block that were missed in
the previous commit.

This update fixes 5 tests in the distribution robustness test suite
and brings us down to 12 failures total.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16599 3a8e244f-b0f2-452b-bcba-4c88e055c3ca