Enabled support for parsing array types in expression contexts so that the following now works:
  def f(e) ...
  def f(type t) ...

  f([1..3] 2); // resolves to first function above
  f([1..3] int); // resolves to second function above

When the parser encounters a forall expression, it now produces a
function that first determines whether the inner expression is a type.
Then, if it is a type, the function returns an array type; otherwise,
the function implements the forall expression.

The parser still parses array types in type contexts.  Next up, we can
unify expressions and types in the parser, removing all remaining
limitations.  For example, we cannot parse a sync or single type in an
expression context today.

Also completed some minor cleanup related to parsing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16612 3a8e244f-b0f2-452b-bcba-4c88e055c3ca