Removed support for the special array declaration syntactic sugar.
Prior to this change, the declaration

  var A: [i in D] int = foo(i);

would be changed to

  var A: [D] int = [i in D] foo(i);

Now, the declaration would result in a compilation error

  error: 'i' undeclared (first use this function)

This syntactic sugar had a number of issues, but it really becomes a
problem when thinking about skyline arrays:

  var A: [i in 1..n] [1..i] int = 3;

This would be translated to

   var A: [i in 1..n] [1..i] int = [i in 1..n] 3;

But this is not what the programmer would want here, and it would lead
to an error.  The issue is that the programmer is presumably using the
index in order to define a skyline array, not to use this syntactic
sugar.  The presumable intent would be for

   var A: [i in 1..n] [1..i] int;
   A = 3;

As is, we would get something similar to

   var A: [i in 1..n] [1..i] int;
   A = [i in 1..n] 3;

which translates to

   forall (a,i) in (A,1..n) do
     a = 3;

and the shapes in the simultaneous iteration don't match.

We can make the syntactic sugar apply to all domains if the first
domain has an index.  This is more complicated and I think fails with
the example:

   var a: [i in 1..n] [1..3] int = (i, i+1, i+2);

Eliminating the syntactic sugar (as is done in this commit) is not a
big loss especially when we get shape-preserving iterators.  In that
case, we can rewrite the original example

  var A: [i in D] int = foo(i);

as

  var A = [i in D] foo(i);

and the resulting type will be an array declared over domain D.  This
will work because forall expressions will be iterators with a shape
defined by their iterator expression (in this case D).  Thus the type
of A is inferred to be "[D] int".  We thus avoid the syntactic sugar
and use Chapel's basic type inference to avoid the extra typing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16615 3a8e244f-b0f2-452b-bcba-4c88e055c3ca