Removed some DSI functions from the domain classes, implementing the same functionality in the wrapper record instead. This reduces the amount of code that needs to be written when creating a new type of distribution.
Removed Dom.dsiStrideBy and Dom.dsiSlice and implemented them in the wrapper
by building a tuple of ranges and creating the new domain with
dsiBuildArithmeticDom. Removed Dom.dsiPosition, implemented in the wrapper with
calls to dsiIndexOrder. Removed non-tuple index versions of Dom.dsiMember,
and Dom.dsiIndexOrder, make the wrapper bundle the index into a tuple.

Added dsiBuildArithmeticDom to Cyclic, BlockCyclic and GUPArithmetic.

Changed the Cyclic array serial writer to follow the same pattern as Default
and Block arrays instead of printing one element per line.

Updated some tests to use indexOrder instead of order and to preserve shape
when printing cyclic arrays.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16627 3a8e244f-b0f2-452b-bcba-4c88e055c3ca