Implemented dsiCreateReindexDist() and dsiReindex() for the Block distribution.  This gives us both reindexing and basic aliasing (currently aliasing is implemented via reindexing).
There was a decision made that you should only be able to use a
non-distributed domain as the reindex domain.  This is reflected in
the test programs (sorry I forgot to put that in the last log
message).  In addition, the reindexed arrays are not linked to the
original reindex domain.  The spec should be updated to note this.

All reindexing and aliasing tests in the distribution robustness suite
are passing for Block.  Only 8 of 56 tests are still failing.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16631 3a8e244f-b0f2-452b-bcba-4c88e055c3ca