Rewrote existing tests using a rank-independent functions as per Steve's suggestion.
Added a test that repeately modifies the reindexed array and the
original.  This test was based on a test that I used to look for a
memory leak that was exposed by reindexing, but it does not test for
the memory leak.

All tests pass for the default and Block distributions.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16638 3a8e244f-b0f2-452b-bcba-4c88e055c3ca