Fixed a bug in function resolution in which functions that are not more visible than other functions could be determined to be more visible because the search through the scopes had numerous flaws.
Removed a future.

Note that, before Sung's change of revision 16641, this change would
result in ambiguities when resolving calls to _computeMyChunk which is
defined in both DefaultArithmetic and ChapelRange.  This is because
this function is called from within a generic method.  Thus both
function definitions can be reached "equally well" from most
instantiation points.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16643 3a8e244f-b0f2-452b-bcba-4c88e055c3ca