Fixed a bug in which we would inappropriately issue a warning that control reaches the end of a function that returns a value (without returning a value) if the function returns a sync or single value via a specified return type.  Removed a future.
Simplified a conditional that guards this warning by applying De
Morgan's law twice.  It's actually readable now, but I also added a
comment.

Discussion:
----------

Should we issue an error if a specified return type is a sync or
single type?

I don't think there is much utility in defining a function that
returns a sync or single type explicitly.  Note that there is no way
to do this with an inferred return type since any sync or single value
would be implicitly read in the return statment.  Moreover, if the
return type is explicitly sync or single, the following still holds:

  (1) if the return expression is a sync or single variable, it will
      be read (and thus a new sync or single will be created and
      returned)

  (2) at any call site, the mere evaluation of this function will
      result in the sync or single being read; if it is assigned to a
      sync or single this is as if a value that were not a sync or
      single were assigned.

That is, in the code

  1  var s: sync int = 0;
  2
  3  def f(): sync int
  4    return s;
  5
  6  s = f();

the variable s is initialized in line 1 making it full.  Then the
variable s is read in line 4 due to the function call in line 6.  This
value is then put into a newly created sync variable (I suspect) and
returned.  Then in line 6, the newly created sync variable is read and
the value is written into s.

Thus I don't think there is any utility in allowing such a specified
return type, and it is somewhat confusing.  Note too that sync and
single variables are passed to functions by reference, making their
return even more complicated if somehow they are not read when
returned.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16650 3a8e244f-b0f2-452b-bcba-4c88e055c3ca