Added a check in the compiler to ensure that iterators are not defined to return types or parameters (functions are allowed to be defined as such).
Added a test of an iterator that is defined to return parameters.

Removed a future related to a test of an iterator that is defined to
return types.

Added a test of a function that returns an array type, but should
return a value.  This test issues an appropriate error.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16653 3a8e244f-b0f2-452b-bcba-4c88e055c3ca