Moved error checking from dsiReindex into _array.reindex().  Also added check to to make sure the reindexing domain is not distributed (default distribution).
Added tests that make sure we catch the error cases.

Updated test_reindexing4.chpl because it was doing an invalid
reindexing by our current definition (i.e., domain distributed).
Possibly we should allow this in the case where the domain is the same
as the one it was declared over, i.e., it's not really a reindexing.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16665 3a8e244f-b0f2-452b-bcba-4c88e055c3ca