Removed two futures for working tests.
The test classes/deitz/test_infer_nil infers a variable to have object
type from nil and then assigns a class instance to this variable.

The test users/shetag/iteratorWithOn tries to use an on-statement in a
serial iterator.  This is only allowed in parallel iterators.  The
error doesn't show up if using COMM=none or --local because the
on-statement is removed in these cases.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16674 3a8e244f-b0f2-452b-bcba-4c88e055c3ca