Added support for eliding the parameter expression and query expression from a variable argument function definition.
Background:
----------

From an e-mail I sent:

"To declare a function that takes a variable number of arguments, we
currently require a parameter expression or a query expression and the
query expression can be anonymous.  So if I want to create a function that
takes exactly p arguments, I can write

   def f(a...p) {

where p is a parameter value.  If I want to take an unknown number of 
arguments and store that in a parameter, I can write

   def f(a...?numArgs) {

If I don't want to name this, I can write

   def f(a...?) {

However, I cannot simply write

   def f(a...) {

to take a variable number of arguments.  The number of arguments can be 
easily determined by writing 'a.size' since a is a tuple.

This seems somewhat inconsistent since we often allow the query expression 
to be omitted altogether if there are no constraints.  The grammar is easy 
to change to allow this last option."


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16677 3a8e244f-b0f2-452b-bcba-4c88e055c3ca