Improved error messages associated with assigning values to types. Fixed a few places in the internal modules where we assigned values to types.  Now we explicitly use .type on the value.  Removed two futures.
Fixed a bug in which the names of instantiated types could contain
extra parentheses if the generic was partially instantiated (due to
generic fields with default values).  Thus C(int,int) will never be
named C(int)(int) depending on instantiation order.  Removed two
futures.

Reworked the implementation of type aliases.  Prior to this change, a
type alias could alias an arbitrary expression and the compiler,
during scope resolution, would replicate this arbitrary expression
everywhere that type alias was used.  There are numerous problems with
this compilation strategy.  Now the type alias is treated as a type
variable and resolved during function resolution.  This is made
possible due to better handling of runtime types such as arrays and
domains.

This new implementation makes it so that we treat type aliases more
similarly to variables in that they need to be used after they are
defined.  For example, we used to allow

  type tt = t;
  type t = int;

but this would now compile to the error "'t' used before defined".
Updated several .goods.

Added a test of assigning a value to a type as in "int = 2".  This
used to compile to the potentially infuriating error message "type
mismatch in assignment from int to int".  It now results in the more
accurate error message "illegal assignment of value to type".

Added a test that creates a local type alias to a sync type.  Other
tests in the suite create a global type alias to a sync type.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16678 3a8e244f-b0f2-452b-bcba-4c88e055c3ca