Fixed a bug in which a use in a block, loop, or any other scoped construct that is top-level to a module, i.e., not in its own function, would be propagated up to that module making more things visible than should be.  Removed a future.
Removed an unnecessary insertion of an empty block statement for every
use statement that was not a use of ChapelStandard.  This code was
inserted/orphaned in revision 15954 when initialization of modules was
decoupled from use statements.  Before this change, a block statement
was inserted, but it wasn't empty.  It called the initialization
function for a module.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16686 3a8e244f-b0f2-452b-bcba-4c88e055c3ca