Added rank changing to block distribution.
Added a method dsiCreateRankChangeDist to distribution classes that creates a
new distribution that distributes indices to the same locales as the original
distribution, but with lower rank.

Moved checkRankChange into ChapelArray so it only needs to be written once
instead of in every distribution.

Moved the implementation of dsiRankChange for domains into ChapelArray and use
dsiCreateRankChangeDist.

Implemented dsiRankChange for block arrays by doing rank changes on the
default arrays representing the local pieces and aliasing them. 

Changed domains/deitz/test_domain_rankchange to use indices that are not out
of bounds for rank changes.

Regression: test/arrays/marybeth/CMO_array.chpl. This test has been passing
lately by getting really lucky with the few features it uses. It doesn't
declare most of the DSI methods. A.domain.dist is nil. I'll look into getting
this test working again soon.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16688 3a8e244f-b0f2-452b-bcba-4c88e055c3ca